# api_test/generated/models/case_input.py
# coding: utf-8

"""
    API Develop Studio

    Studio management API. Existing business validation and error contracts are retained.

    The version of the OpenAPI document: 1.0.1
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
from typing import Any, Dict, List, Optional, Union
from pydantic import BaseModel, ConfigDict, Field, field_validator, StrictBool, StrictFloat, StrictInt, StrictStr


from api_test.generated.models.case_request import CaseRequest
from api_test.generated.models.expected_response import ExpectedResponse
from api_test.generated.models.storage_metadata import StorageMetadata
from api_test.generated.models.variables_input import VariablesInput

class CaseInput(BaseModel):
    """Generated contract DTO. Additional document fields and explicit nulls are retained."""
    model_config = ConfigDict(extra='allow', populate_by_name=True, protected_namespaces=())
    name: Optional[StrictStr] = None
    project: Optional[StrictStr] = None
    base_url_name: Optional[StrictStr] = None
    timeout: Optional[Union[StrictFloat, StrictInt]] = None
    request: CaseRequest
    expected: ExpectedResponse
    storage: Optional[StorageMetadata] = Field(default=None, alias="_storage")
    expected_body_raw: Optional[StrictStr] = Field(default=None, alias="_expectedBodyRaw")
    variables: Optional[VariablesInput] = None


    def to_dict(self):
        return self.model_dump(by_alias=True, exclude_unset=True)

    @classmethod
    def from_dict(cls, value):
        return cls.model_validate(value)