# openapi/studio.yaml
openapi: 3.0.3
info:
  title: API Develop Studio
  version: 1.0.1
  description: Studio management API. Existing business validation and error contracts
    are retained.
paths:
...
  /api/cases:
    get:
      operationId: list_cases
      tags:
      - cases
      summary: list cases
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceList'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-studio-handler: documents.list_documents
      x-studio-parts: '[''cases'']'
      x-studio-route: /api/cases
      x-studio-upload: false
      x-studio-alias: true
      parameters:
      - name: project
        in: query
        required: false
        schema: *id001
  /api/cases/{reference}/revisions:
    get:
      operationId: list_cases_revisions
      tags:
      - cases
      summary: list cases revisions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionList'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-studio-handler: documents.list_revisions
      x-studio-parts: '[''cases'']'
      x-studio-route: /api/cases/{reference:path}/revisions
      x-studio-upload: false
      x-studio-alias: true
      parameters:
      - name: reference
        in: path
        required: true
        schema: *id001
        description: Slash-separated document or attachment reference.
  /api/cases/{reference}:
    get:
      operationId: get_cases
      tags:
      - cases
      summary: get cases
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseView'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-studio-handler: documents.get_document
      x-studio-parts: '[''cases'']'
      x-studio-route: /api/cases/{reference:path}
      x-studio-upload: false
      x-studio-alias: false
      parameters:
      - name: reference
        in: path
        required: true
        schema: *id001
        description: Slash-separated document or attachment reference.
    put:
      operationId: save_cases
      tags:
      - cases
      summary: save cases
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-studio-handler: documents.save_document
      x-studio-parts: '[''cases'']'
      x-studio-route: /api/cases/{reference:path}
      x-studio-upload: false
      x-studio-alias: false
      parameters:
      - name: reference
        in: path
        required: true
        schema: *id001
        description: Slash-separated document or attachment reference.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseInput'
    delete:
      operationId: delete_cases
      tags:
      - cases
      summary: delete cases
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-studio-handler: documents.delete_document
      x-studio-parts: '[''cases'']'
      x-studio-route: /api/cases/{reference:path}
      x-studio-upload: false
      x-studio-alias: false
      parameters:
      - name: reference
        in: path
        required: true
        schema: *id001
        description: Slash-separated document or attachment reference.
...
  /example-api:
    get:
      operationId: get_example
      tags:
      - example
      summary: get example
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: Example user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '401':
          description: Missing or invalid example API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '404':
          description: Example endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
      x-studio-handler: example.handle_example
      x-studio-parts: '[]'
      x-studio-route: /example-api
      x-studio-upload: false
      x-studio-alias: true
    post:
      operationId: post_example
      tags:
      - example
      summary: post example
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: Example user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '401':
          description: Missing or invalid example API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '404':
          description: Example endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
      x-studio-handler: example.handle_example
      x-studio-parts: '[]'
      x-studio-route: /example-api
      x-studio-upload: false
      x-studio-alias: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExampleInput'
  /example-api/{reference}:
    get:
      operationId: get_example_resource
      tags:
      - example
      summary: get example resource
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: Example user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '401':
          description: Missing or invalid example API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '404':
          description: Example endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
      x-studio-handler: example.handle_example
      x-studio-parts: '[]'
      x-studio-route: /example-api/{reference:path}
      x-studio-upload: false
      x-studio-alias: false
      parameters:
      - name: reference
        in: path
        required: true
        schema: *id001
        description: Slash-separated document or attachment reference.
    post:
      operationId: post_example_resource
      tags:
      - example
      summary: post example resource
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: Example user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '401':
          description: Missing or invalid example API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '404':
          description: Example endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
      x-studio-handler: example.handle_example
      x-studio-parts: '[]'
      x-studio-route: /example-api/{reference:path}
      x-studio-upload: false
      x-studio-alias: false
      parameters:
      - name: reference
        in: path
        required: true
        schema: *id001
        description: Slash-separated document or attachment reference.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExampleInput'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: *id001
        code: *id001
        currentRevision: &id002
          type: integer
        runId: *id001
      required:
      - error
    StorageMetadata:
      type: object
      properties:
        id:
          type: string
        revision: *id002
        created_by: *id001
        updated_by: *id001
        created_at: *id001
        updated_at: *id001
        content_hash: *id001
        projectionPending:
          type: boolean
    SecretInput:
      type: object
      properties:
        value:
          type: string
          writeOnly: true
        preserve: &id003
          type: boolean
    SecretState:
      type: object
      properties:
        configured: *id003
        value:
          type: string
          description: Only the explicitly public bundled Example fixture.
      required:
      - configured
    VariablesInput:
      type: object
      properties:
        plain:
          type: object
          additionalProperties: *id001
        secret:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SecretInput'
    VariablesView:
      type: object
      properties:
        plain:
          type: object
          additionalProperties: *id001
        secret:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SecretState'
...
    FormDataItem:
      type: object
      properties:
        key: *id001
        value: *id001
        file: *id001
        filename: *id001
        content_type: *id001
      required:
      - key
    CaseRequest:
      type: object
      properties:
        method: *id001
        url: *id001
        headers:
          type: object
          additionalProperties: *id001
        auth: *id010
        auth_profile: *id001
        body: {}
        text: *id001
        form_data:
          type: array
          items:
            $ref: '#/components/schemas/FormDataItem'
      required:
      - url
    ExpectedResponse:
      type: object
      properties:
        status: *id002
        body: {}
        strict: *id003
        max_response_time_ms: &id011
          type: number
        assertions:
          type: array
          items: *id010
        validation_modes:
          type: object
          additionalProperties: *id003
    CaseInput:
      type: object
      properties:
        name: *id001
        project: *id001
        base_url_name: *id001
        timeout: *id011
        request: &id012
          $ref: '#/components/schemas/CaseRequest'
        expected: &id013
          $ref: '#/components/schemas/ExpectedResponse'
        _storage: &id014
          $ref: '#/components/schemas/StorageMetadata'
        _expectedBodyRaw: *id001
        variables:
          $ref: '#/components/schemas/VariablesInput'
      required:
      - request
      - expected
    CaseView:
      type: object
      properties:
        name: *id001
        project: *id001
        base_url_name: *id001
        timeout: *id011
        request: *id012
        expected: *id013
        _storage: *id014
        _expectedBodyRaw: *id001
        variables:
          $ref: '#/components/schemas/VariablesView'
      required:
      - request
      - expected
...
    ReferenceList:
      type: object
      properties:
        items:
          type: array
          items: *id001
        details:
          type: object
          additionalProperties: *id010
      required:
      - items
    RevisionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StorageMetadata'
      required:
      - items
    SaveResponse:
      type: object
      properties:
        path: *id001
        _storage:
          $ref: '#/components/schemas/StorageMetadata'
      required:
      - path
      - _storage
    DeleteResponse:
      type: object
      properties:
        deleted: *id001
        deleted_pipelines:
          type: array
          items: *id001
      required:
      - deleted
...
    ExampleInput:
      type: object
      properties:
        name: *id001
    ExampleResponse:
      type: object
      properties:
        status: *id001
        id: *id002
        name: *id001
        authorized: *id003
        message: *id001
        error: *id001